Rearrange test helpers
